test 12

+++ custom_ui.py
@@ -199,8 +199,8 @@ class Window(QMainWindow, Ui_MainWindow):
         # print(self.x()+ 650, self.y())
         parent_geom = self.geometry()
         dialog_width, dialog_height = 254, 162
-        x_pos = parent_geom.x() + (parent_geom.width() - dialog_width) / 2
-        y_pos = parent_geom.y() + (parent_geom.height() - dialog_height) / 2
+        x_pos = int(parent_geom.x() + (parent_geom.width() - dialog_width) / 2)
+        y_pos = int(parent_geom.y() + (parent_geom.height() - dialog_height) / 2)
         self.dialog.setGeometry(x_pos, y_pos, dialog_width, dialog_height)
         # self.dialog.setGeometry(650, 0, 254, 162) # - 228 + 58
         # x_pos = self.x() + 650
